change integer to big integer

+++ backend/app/models/user.py
@@ -1,12 +1,11 @@
 # app/models/user.py
-from sqlalchemy import Column, Integer, String, Boolean
-# Импортируем Base из нашего файла настроек БД (который мы исправили ранее)
+from sqlalchemy import Column, Integer, String, Boolean, BigInteger
 from app.database import Base 
 
 class User(Base):
     __tablename__ = "users"
 
-    id = Column(Integer, primary_key=True, index=True)
+    id = Column(BigInteger, primary_key=True, index=True)
     email = Column(String, unique=True, index=True)
     # ВАЖНО: В базе мы храним хэш пароля, а не сам пароль!
     hashed_password = Column(String)
